Refined resilience in face of BLE failure.
Now, if we detect that the BLE stack is borked, we signal the need
to reset it and wait until we get killed.

Otherwise, we assume the BLE stack is good to go.

We don't actually do the reset ourselves, since we need to take
the whole BLE stack down to do that.